fix(extension): support dynamic fractional units in scrape and pool
Parse theoretical/practical units as floats to handle 0.5, 0.25, 0.75 etc
from amoozeshyar (e.g. تربیت بدنی 0.50+0.50). Add toFloat helpers with
Persian digit and decimal separator handling.

fix(registry): allow dynamic fractional units in schemas

Remove multipleOf 0.5 constraint from offering and chart schemas
so any decimal 0-20 is valid and chart exports preserve 0.25 etc.

+++ packages/registry/src/schema/chart.ts
@@ -16,8 +16,8 @@ export type ChartSemester = z.infer<typeof chartSemesterSchema>
  */
 export const chartCourseSchema = z.object({
   name: z.string().min(1).max(255),
-  theoreticalUnits: z.number().int().min(0).max(20),
-  practicalUnits: z.number().int().min(0).max(20),
+  theoreticalUnits: z.number().min(0).max(20),
+  practicalUnits: z.number().min(0).max(20),
   /** Prerequisites: course NAMES (string[]) or required passed units (number, e.g. 100) or required terms (e.g. { term: 5 } for گذراندن 5 نیمسال). */
   prerequisites: z
     .union([
